+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 🚀 Features
-- None
+- [QQ -> Discord] 当 QQ 群内的文件大小超过上限后将会发送该文件的直链（该链接可能会过期，**且该功能仅适用于支持 go-cqhttp 扩展 API 的协议端**）
 
 ### 🐞 Bug Fixes
-- [Discord -> QQ] 修复了音频文件无法通过 bridge 的问题
+- None